requirements.txt
cookiecutter

# Testing
docker-py
snakemake
pytest
pytest-runner
pep8
pytest-cookies
pyflakes
